Move implementation to its own provider
