+++ src/Umbraco.Cms.Api.Management/Factories/IDocumentUrlFactory.cs
@@ -6,5 +6,6 @@ namespace Umbraco.Cms.Api.Management.Factories;
 public interface IDocumentUrlFactory
 {
     Task<IEnumerable<DocumentUrlInfo>> CreateUrlsAsync(IContent content);
+
     Task<IEnumerable<DocumentUrlInfoResponseModel>> CreateUrlSetsAsync(IEnumerable<IContent> contentItems);
 }
